backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-environ==0.9.0 # This
psycopg2-binary==2.9.6
sqlparse==0.4.3
djangorestframework==3.14.0